Separate Create and Join components from Menu
